requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.110.0
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
idna==3.10
protobuf==5.29.4
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.0.1
redis==5.2.1
requests==2.31.0
setuptools==78.1.0
sniffio==1.3.1
starlette==0.36.3
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.29.0